+++ gulpfile.js
@@ -11,18 +11,18 @@ var uglify = require('gulp-uglify');
 var usemin = require('gulp-usemin');
 var rimraf = require('rimraf');
 var rename = require('gulp-rename');
+var merge = require('merge2');
 
 var tsProject = ts.createProject({
-    declarationFiles: false,
+    declaration: true,
     noExternalResolve: true,
     noImplicitAny: false,
     removeComments: false,
-    sortOutput: true,
     target: 'ES5',
     emitDecoratorMetadata: false
 });
 
-gulp.task('typescript', ['clean'], function() {
+gulp.task('typescript', ['clean'], function () {
     var tsResult = gulp.src(['src/*.module.ts', 'src/*.ts', 'typings/**/*.d.ts'])
         .pipe(sourcemaps.init())
         .pipe(ts(tsProject));
@@ -35,12 +35,18 @@ gulp.task('typescript', ['clean'], function() {
         .pipe(sourcemaps.write('./'))
         .pipe(gulp.dest('dist/'));
 
-    return jsResult;
+	var dtsResult = tsResult.dts
+		.pipe(gulp.dest('dist/'));
+
+	return merge([
+       	jsResult,
+        dtsResult
+    ]);
 });
 
-gulp.task('usemin', ['typescript'], function() {
+gulp.task('usemin', ['typescript'], function () {
     return gulp.src('dist/*.js')
-        .pipe(sourcemaps.init({loadMaps: true}))
+        .pipe(sourcemaps.init({ loadMaps: true }))
         .pipe(uglify())
         .pipe(rename({
             suffix: '.min'
@@ -49,15 +55,10 @@ gulp.task('usemin', ['typescript'], function() {
         .pipe(gulp.dest('dist'));
 });
 
-gulp.task('copy-typings', ['usemin'], function () {
-    return gulp.src('src/cloudinary-angular.d.ts')
-        .pipe(gulp.dest('dist'));
-});
-
-gulp.task('clean', function(cb) {
+gulp.task('clean', function (cb) {
     rimraf('dist', cb);
 });
 
-gulp.task('default', function() {
-    gulp.start('copy-typings');
+gulp.task('default', function () {
+    gulp.start('usemin');
 });