+++ src/cloudinary.module.ts
@@ -1,3 +1,0 @@
-namespace cloudinary {
-	const appModule = angular.module('cloudinary', []);
-}